refactor(axon-vectors): move Arc<dyn VectorStore> blanket impl to store/arc_forward.rs
Keeps store.rs within the 500-line monolith budget (was pushed to 520 by the
blanket forwarding impl; now 471). Pure code move + super::Result import.
